+++ modules/drivers/rfid/rfid_component.cc
@@ -73,7 +73,7 @@ void RfidComponent::Action() {
       if (count == 10) {
         AINFO << "origin id from buffer[10]: " << buffer[10];
         // uint32_t station_id = buffer[10] - '0';
-        uint32_t station_id = buf;
+        uint8_t station_id = buf;
         AINFO << "TRANSFER ID :" << station_id;
 
         apollo::drivers::RFID rfid;
